chg: dev: code reworked for new_resolution and reusability improved
